role-invite-bot/requirements.txt
# 🤖 МИНИМАЛЬНЫЕ зависимости для ультра-простого бота
discord.py==2.3.2
python-dotenv==1.0.1
